backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
httpx==0.26.0
spacy==3.7.2
pydantic==2.5.3
google-generativeai==0.3.2
